final version of project
This is the exact version of the project my history teacher received, with the exception of his name being removed.
Although I don't plan to do so, if I update my history teacher's version of the project, I will update this one as well.  Other than that, this repo just serves as an archive.